requirements.txt
certifi==2022.12.7
elastic-transport==8.4.0
elasticsearch==8.6.2
pyirsdk==1.3.2
PyYAML==6.0
urllib3==1.26.14
